fixed pathing around interactable items
Interactables now spawn on "water" so the player paths around them. We may want to remove their gameobject, or allow movement to be cancelled, so the player doesn't get stuck against it

+++ Assets/Scripts/HexagonHelpers.cs
@@ -127,7 +127,7 @@ public class HexagonHelpers : MonoBehaviour
         return path;
     }
 
-    public static void SetPosition(GameObject obj, int indxX, int y, int indxZ)
+    public static void SetPosition(GameObject obj, int indxX, float y, int indxZ)
     {
         float iVal = indxX;
         if (indxZ % 2 == 1)
